design: PlaceholderView 각각 분리
#82
