+++ Bridge/Sources/Data/Network/DTOs/Main/ProjectDTO.swift
@@ -42,96 +42,96 @@ extension ProjectDTO {
 
 extension ProjectDTO {
     static var projectTestArray: [ProjectDTO] = [
-        ProjectDTO(
-            id: "1",
-            title: "모임 플랫폼 디자이너 구합니다",
-            overview: "모임 플랫폼",
-            dueDate: Date(),
-            startDate: Date(),
-            endDate: Date(),
-            recruit: [],
-            tagLimit: [],
-            meetingWay: "온라인",
-            stage: "기획 중",
-            userEmail: ""
-        ),
-        ProjectDTO(
-            id: "2",
-            title: "웹 사이트 디자이너 구해요!!",
-            overview: "모임 플랫폼",
-            dueDate: Date(),
-            startDate: Date(),
-            endDate: Date(),
-            recruit: [],
-            tagLimit: [],
-            meetingWay: "온라인",
-            stage: "기획 중",
-            userEmail: ""
-        ),
-        ProjectDTO(
-            id: "3",
-            title: "웹 사이트 디자이너 구해요!!",
-            overview: "모임 플랫폼",
-            dueDate: Date(),
-            startDate: Date(),
-            endDate: Date(),
-            recruit: [],
-            tagLimit: [],
-            meetingWay: "온라인",
-            stage: "기획 중",
-            userEmail: ""
-        ),
-        ProjectDTO(
-            id: "4",
-            title: "웹 사이트 디자이너 구해요!!",
-            overview: "모임 플랫폼",
-            dueDate: Date(),
-            startDate: Date(),
-            endDate: Date(),
-            recruit: [],
-            tagLimit: [],
-            meetingWay: "온라인",
-            stage: "기획 중",
-            userEmail: ""
-        ),
-        ProjectDTO(
-            id: "5",
-            title: "웹 사이트 디자이너 구해요!!",
-            overview: "모임 플랫폼",
-            dueDate: Date(),
-            startDate: Date(),
-            endDate: Date(),
-            recruit: [],
-            tagLimit: [],
-            meetingWay: "온라인",
-            stage: "기획 중",
-            userEmail: ""
-        ),
-        ProjectDTO(
-            id: "6",
-            title: "웹 사이트 디자이너 구해요!!",
-            overview: "모임 플랫폼",
-            dueDate: Date(),
-            startDate: Date(),
-            endDate: Date(),
-            recruit: [],
-            tagLimit: [],
-            meetingWay: "온라인",
-            stage: "기획 중",
-            userEmail: ""
-        ),
-        ProjectDTO(
-            id: "7",
-            title: "웹 사이트 디자이너 구해요!!",
-            overview: "모임 플랫폼",
-            dueDate: Date(),
-            startDate: Date(),
-            endDate: Date(),
-            recruit: [],
-            tagLimit: [],
-            meetingWay: "온라인",
-            stage: "기획 중",
-            userEmail: ""
-        )
+//        ProjectDTO(
+//            id: "1",
+//            title: "모임 플랫폼 디자이너 구합니다",
+//            overview: "모임 플랫폼",
+//            dueDate: Date(),
+//            startDate: Date(),
+//            endDate: Date(),
+//            recruit: [],
+//            tagLimit: [],
+//            meetingWay: "온라인",
+//            stage: "기획 중",
+//            userEmail: ""
+//        ),
+//        ProjectDTO(
+//            id: "2",
+//            title: "웹 사이트 디자이너 구해요!!",
+//            overview: "모임 플랫폼",
+//            dueDate: Date(),
+//            startDate: Date(),
+//            endDate: Date(),
+//            recruit: [],
+//            tagLimit: [],
+//            meetingWay: "온라인",
+//            stage: "기획 중",
+//            userEmail: ""
+//        ),
+//        ProjectDTO(
+//            id: "3",
+//            title: "웹 사이트 디자이너 구해요!!",
+//            overview: "모임 플랫폼",
+//            dueDate: Date(),
+//            startDate: Date(),
+//            endDate: Date(),
+//            recruit: [],
+//            tagLimit: [],
+//            meetingWay: "온라인",
+//            stage: "기획 중",
+//            userEmail: ""
+//        ),
+//        ProjectDTO(
+//            id: "4",
+//            title: "웹 사이트 디자이너 구해요!!",
+//            overview: "모임 플랫폼",
+//            dueDate: Date(),
+//            startDate: Date(),
+//            endDate: Date(),
+//            recruit: [],
+//            tagLimit: [],
+//            meetingWay: "온라인",
+//            stage: "기획 중",
+//            userEmail: ""
+//        ),
+//        ProjectDTO(
+//            id: "5",
+//            title: "웹 사이트 디자이너 구해요!!",
+//            overview: "모임 플랫폼",
+//            dueDate: Date(),
+//            startDate: Date(),
+//            endDate: Date(),
+//            recruit: [],
+//            tagLimit: [],
+//            meetingWay: "온라인",
+//            stage: "기획 중",
+//            userEmail: ""
+//        ),
+//        ProjectDTO(
+//            id: "6",
+//            title: "웹 사이트 디자이너 구해요!!",
+//            overview: "모임 플랫폼",
+//            dueDate: Date(),
+//            startDate: Date(),
+//            endDate: Date(),
+//            recruit: [],
+//            tagLimit: [],
+//            meetingWay: "온라인",
+//            stage: "기획 중",
+//            userEmail: ""
+//        ),
+//        ProjectDTO(
+//            id: "7",
+//            title: "웹 사이트 디자이너 구해요!!",
+//            overview: "모임 플랫폼",
+//            dueDate: Date(),
+//            startDate: Date(),
+//            endDate: Date(),
+//            recruit: [],
+//            tagLimit: [],
+//            meetingWay: "온라인",
+//            stage: "기획 중",
+//            userEmail: ""
+//        )
     ]
 }
